# business-core/docs/api/business-core-openapi.yaml
openapi: 3.1.0
info:
  title: Business Core API
  version: "1.0.0"
  description: |
    API du Business Core — noyau métier générique intégré au kernel RT-Comops.

    Le Business Core expose un modèle de métier piloté par les métadonnées :
    le développeur déclare des Types Métier (le modèle), crée des Entreprises
    (les instances), et déclenche des Opérations. Le Business Core orchestre
    le kernel en façade.

    ## Authentification
    Chaque requête porte la clé Business Core du développeur via les en-têtes
    `X-BC-Client-Id` et `X-BC-Api-Key`. Le secret kernel n'est jamais exposé.

    ## Conventions
    - Versionnement de l'API dans l'URL (`/v1`), distinct de la version d'un Type Métier.
    - Opérations synchrones par défaut (`200`), différées en `202` avec une trace de suivi.
    - Erreurs au format RFC 7807 (Problem Details), enrichies de champs métier.
  contact:
    name: Équipe Business Core
  license:
    name: Usage académique
servers:
  - url: https://business-core.example.com
    description: Serveur du Business Core
security:
  - BusinessCoreKey: []
...
paths:
...
  /v1/business-types/{typeId}/versions/{versionNumber}/roles:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
    post:
      tags: [Contenu de version]
      summary: Déclarer un rôle métier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRole'
      responses:
        '201':
          description: Rôle déclaré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
    get:
      tags: [Contenu de version]
      summary: Lister les rôles métier
      responses:
        '200':
          description: Liste des rôles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
  /v1/business-types/{typeId}/versions/{versionNumber}/roles/{roleId}:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
      - name: roleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags: [Contenu de version]
      summary: Modifier un rôle métier
      description: Met à jour le code ; la catégorie est immuable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRole'
      responses:
        '200':
          description: Rôle mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Contenu de version]
      summary: Supprimer un rôle métier
      description: Refusé (409) si des acteurs actifs le référencent.
      responses:
        '204':
          description: Rôle supprimé
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /v1/businesses:
    post:
      tags: [Entreprises]
      summary: Créer une entreprise
      description: Instancie un Type Métier à une version donnée. L'entreprise reste épinglée à cette version.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusiness'
      responses:
        '201':
          description: Entreprise créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '422':
          $ref: '#/components/responses/Problem'
    get:
      tags: [Entreprises]
      summary: Lister les entreprises
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBusiness'
  /v1/businesses/{businessId}:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    get:
      tags: [Entreprises]
      summary: Consulter une entreprise
      responses:
        '200':
          description: L'entreprise
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Entreprises]
      summary: Modifier une entreprise
      description: Met à jour les métadonnées locales (nom). Pas de rename de l'organisation kernel.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBusiness'
      responses:
        '200':
          description: Entreprise mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Entreprises]
      summary: Archiver une entreprise
      description: Passe le cycle de vie en FERMEE (local + kernel close). Pas de suppression dure.
      responses:
        '204':
          description: Entreprise archivée
        '404':
          $ref: '#/components/responses/Problem'
  /v1/businesses/{businessId}/approve:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    post:
      tags: [Entreprises]
      summary: Approuver l'organisation kernel
      description: |
        Première approbation de gouvernance (`POST /api/organizations/{id}/approve`).
        Distinct de `PUT .../lifecycle` avec ACTIVE (qui appelle `reopen`).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveBusiness'
      responses:
        '200':
          description: Entreprise approuvée (cycleVie ACTIVE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '404':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /v1/businesses/{businessId}/lifecycle:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    put:
      tags: [Entreprises]
      summary: Changer le cycle de vie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleChange'
      responses:
        '200':
          description: Cycle de vie mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
  /v1/businesses/{businessId}/actors:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    post:
      tags: [Entreprises]
      summary: Rattacher un acteur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachActor'
      responses:
        '201':
          description: Acteur rattaché
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
    get:
      tags: [Entreprises]
      summary: Lister les acteurs
      responses:
        '200':
          description: Liste des acteurs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Actor'
  /v1/businesses/{businessId}/actors/{actorId}:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
      - $ref: '#/components/parameters/ActorId'
    get:
      tags: [Entreprises]
      summary: Consulter un acteur
      responses:
        '200':
          description: L'acteur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Entreprises]
      summary: Modifier le rôle d'un acteur
      description: |
        Détache l'association courante et en crée une nouvelle avec le rôle fourni.
        La catégorie (OPERATEUR / BENEFICIAIRE) ne peut pas changer (RG-04).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActor'
      responses:
        '200':
          description: Nouvelle association acteur
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Actor'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Entreprises]
      summary: Détacher un acteur
      description: Met fin à la période de validité du rattachement.
      responses:
        '204':
          description: Acteur détaché
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    BusinessCoreKey:
      type: apiKey
      in: header
      name: X-BC-Api-Key
      description: Clé d'API du développeur. Accompagnée de l'en-tête X-BC-Client-Id.
  parameters:
    TypeId:
      name: typeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionNumber:
      name: versionNumber
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    BusinessId:
      name: businessId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ActorId:
      name: actorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
    Problem:
      description: Erreur au format RFC 7807
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: Détail d'erreur RFC 7807, enrichi de champs métier.
      properties:
        type:
          type: string
          example: /errors/regle-violee
        title:
          type: string
          example: Règle métier non respectée
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: Un document ordonnance est requis pour cette vente.
        instance:
          type: string
          example: /v1/businesses/abc/operations/vente:execute
        violatedRule:
          type: string
          description: Code de la règle violée (extension métier)
        requiredAction:
          type: string
          enum: [BLOQUER, EXIGER, VALIDER, AJUSTER, ALERTER, DEROGER]
        requiredDocument:
          type: string
...
    CreateRole:
      type: object
      required: [code, categorie]
      properties:
        code:
          type: string
          example: pharmacien-responsable
        categorie:
          type: string
          enum: [OPERATEUR, BENEFICIAIRE]
    UpdateRole:
      type: object
      required: [code]
      properties:
        code:
          type: string
          example: pharmacien-responsable
    Role:
      allOf:
        - $ref: '#/components/schemas/CreateRole'
        - type: object
          properties:
            id:
              type: string
              format: uuid
...
    CreateBusiness:
      type: object
      required: [typeId, versionNumber, nom]
      properties:
        typeId:
          type: string
          format: uuid
        versionNumber:
          type: integer
        nom:
          type: string
          example: Pharmacie du Centre
        organizationId:
          type: string
          format: uuid
          description: Référence Organization kernel (créée si absente)
    UpdateBusiness:
      type: object
      required: [nom]
      properties:
        nom:
          type: string
          example: Pharmacie du Centre
    ApproveBusiness:
      type: object
      properties:
        reason:
          type: string
          description: Motif d'approbation (défaut « Approbation initiale »)
          example: Validation KYC
    Business:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nom:
          type: string
        typeId:
          type: string
          format: uuid
        versionNumber:
          type: integer
        organizationId:
          type: string
          format: uuid
        cycleVie:
          type: string
          enum: [ACTIVE, SUSPENDUE, FERMEE]
    LifecycleChange:
      type: object
      required: [cycleVie]
      properties:
        cycleVie:
          type: string
          enum: [ACTIVE, SUSPENDUE, FERMEE]
    AttachActor:
      type: object
      required: [roleCode, acteurKernelId]
      properties:
        roleCode:
          type: string
        acteurKernelId:
          type: string
          format: uuid
          description: Référence Actor (opérateur) ou Tiers (bénéficiaire) kernel
        valideDepuis:
          type: string
          format: date
    UpdateActor:
      type: object
      required: [roleMetierId]
      properties:
        roleMetierId:
          type: string
          format: uuid
          description: Nouveau rôle métier (même catégorie que l'actuel)
    Actor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        roleCode:
          type: string
        categorie:
          type: string
          enum: [OPERATEUR, BENEFICIAIRE]
        acteurKernelId:
          type: string
          format: uuid
        valideDepuis:
          type: string
          format: date
        valideJusqua:
          type: string
          format: date
...
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
...
    PageBusiness:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Business'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...